requirements.txt
# Backend & API
fastapi
uvicorn
python-dotenv
pydantic>=2.0.0
python-multipart

# AI, LLM & RAG
openai
langchain
langchain-openai
langchain-community
langchain-chroma
langchain-core
chromadb
tiktoken

# Data Processing & Visualization
pandas
numpy
plotly

# Frontend
streamlit
requests
python-multipart>=0.0.7
